Revert "refactor(labs): Roll our own ThemeProvider"
This reverts commit 778eaabb37116f2bc4b7a2464075346d0ee2b118.
