+++ modules/_labs/core/react/lib/CanvasProvider.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
+import {ThemeProvider} from 'emotion-theming';
 import {InputProvider} from '@workday/canvas-kit-react-core';
 import {CanvasTheme} from './theming/types';
-import {ThemeProvider, defaultCanvasTheme} from './theming/theme';
+import {defaultCanvasTheme} from './theming/theme';
 
 export interface CanvasProviderProps {
   theme: CanvasTheme;
@@ -16,7 +17,7 @@ export default class CanvasProvider extends React.Component<CanvasProviderProps>
     const {children, theme} = this.props;
 
     return (
-      <ThemeProvider value={theme}>
+      <ThemeProvider theme={theme}>
         <InputProvider>{children}</InputProvider>
       </ThemeProvider>
     );
