+++ modules/_labs/core/react/stories/stories_theme.tsx
@@ -4,8 +4,8 @@ import {jsx} from '@emotion/core';
 import styled from '@emotion/styled';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-
-import {CanvasTheme, CanvasThemePalette, useTheme} from '../index';
+import useTheme from '../lib/theming/useTheme';
+import {CanvasTheme, CanvasThemePalette} from '../lib/theming/types';
 import README from '../lib/theming/README.md';
 import {H1, colors, type, spacing, borderRadius} from '@workday/canvas-kit-react-core';
